little change of mypcb.h

+++ ch02/mykernel/mypcb.h
@@ -12,17 +12,21 @@
 
  /* CPU-specific state of this task */
 struct Thread {
-    unsigned int  ip;
-    unsigned int  sp;
+    /*
+     * since we will build 32-bit kernel (which we already specify in `menuconfig`),
+     * we also can `unsigned int` which is 32-bit too.
+     */
+    unsigned long  ip; // 32-bit = 4-byte
+    unsigned long  sp;
 };
 
 typedef struct PCB{
     int pid;
-    volatile int state;    /* -1 unrunnable, 0 runnable, >0 stopped */
-    unsigned int stack[KERNEL_STACK_SIZE];
+    volatile long state;    /* -1 unrunnable, 0 runnable, >0 stopped */
+    unsigned long stack[KERNEL_STACK_SIZE];
     /* CPU-specific state of this task */
     struct  Thread thread;
-    unsigned int   task_entry;
+    unsigned long   task_entry;
     struct  PCB *next;
 } tPCB;
